Add pattern based error/warning detection
Currently we use item.type to identify an item as an error / warning. However,
this value isn't always populated. As a result, compilation often results with
all marks set to "info".

To fix this we add regex patterns that can identify the text itself as an error
/ warning.

These patterns are configurable, and default to '^ *error:' and '^ *warning:'
which work for g++ and clang.

NOTE: this can cause false-positive with, e.g., grep if the resulting line
starts with 'error:' / 'warning:'. If that happens, the grep result will be
marked as an error/warning instead of info.